pruebas unitarias crear pedido

+++ microservicio/dominio/src/main/java/com/ceiba/pedido/modelo/dto/DtoPedido.java
@@ -15,8 +15,8 @@ public class DtoPedido {
     private Long comprador;
     private String ciudad;
     private String destino;
-    private Integer flete;
-    private Integer precio;
-    private Integer descuento;
+    private int flete;
+    private int precio;
+    private double descuento;
     private double total;
 }
